requirements.txt
fastapi
uvicorn[standard]

langchain-core
langchain-groq

python-dotenv
requests